Integrate EventCache with FullCalendar view layer (#303)
This PR constitutes a rewrite of most of the core logic in the plugin. Event and Source "models" are replaced with Calendar classes that encapsulate all the logic around parsing events and, in the case of editable Full Note and Daily Note calendars, writing updates back into the filesystem. These Calendars interface with the EventCache introduced in 4ce7855c to provide a persistent collection of all events viewable in the calendar.