Show tasks without a time assigned in the all-day section
This assigns a default startTime to daily note tasks without inline
attributes. Which allows creating a list of tasks and then dragging them
to an appropriate time.

Note that globalAttrs is cast as any because the inferred zod type is
missing all the fields from the discriminated union (please let me know
if this is expected / what the fix is).